Print an integer number

+++ 0x04-more_functions_nested_loops/101-print_number.c
@@ -7,8 +7,8 @@
 void print_number(int n)
 {
 	int b, i, f, t;
-	long int num;
-	long int r;
+	unsigned int num;
+	unsigned int r;
 
 	f = 1;
 	b = 1;
